more

+++ Win32/Kernel.cs
@@ -6,7 +6,7 @@ using System;
 public static class Kernel {
     private const string kernel32 = nameof(kernel32) + ".dll";
     [DllImport(kernel32, CallingConvention = CallingConvention.Winapi)]
-    public extern static ulong GetLastError ();
+    public extern static uint GetLastError ();
     [DllImport(kernel32, SetLastError = true, CallingConvention = CallingConvention.Winapi)]
     public extern static IntPtr GetProcAddress (IntPtr module, [MarshalAs(UnmanagedType.LPStr)] string name);
     [DllImport(kernel32, SetLastError = true, CallingConvention = CallingConvention.Winapi)]
